added manual url specification option

+++ R/download-series-files.R
@@ -51,22 +51,25 @@ geograbr.download.series.files <- function(path=".", gpl = "GPL13534", gses) {
 #' @param path path for saving downloaded GEO series matrix files
 #' @param gse character of gse of interest  
 #' @param sup.file.name name of specific supplemental file from gse to downloads 
+#' @param url optional manual url supply for non-ftp based file downloads
 #'
 #' @export
-geograbr.download.supplementary.file <- function(path=".", gse, sup.file.name) {
+geograbr.download.supplementary.file <- function(path=".", gse, sup.file.name=NULL, url=NULL) {
     require(RCurl)
+    if (is.null(url)){
         prefix <- substring(gse,1,5)
         url <- paste("ftp://ftp.ncbi.nlm.nih.gov/geo/series/",
                       prefix, "nnn/",
                       gse, "/suppl/",
                       sup.file.name, sep="")
-        
+        destination <- file.path(path, basename(url))
+      } else {
+        destination <- file.path(path, sup.file.name)
+      }
         base::cat(base::date(), gse, url[1], "\n")
 
-        destination <- file.path(path, basename(url))
         if (file.exists(destination)) return(destination)
     
-        destination <- file.path(path, basename(url))
         tryCatch(download.file(url, destfile=destination),
                  error=function(e) {
                      unlink(destination)
@@ -77,3 +80,4 @@ geograbr.download.supplementary.file <- function(path=".", gse, sup.file.name) {
             return(destination)
     NA
 }
+
